requirements.txt
streamlit
pandas
pytz
pytesseract
st-gsheets-connection
plotly
opencv-python-headless
fpdf
google-generativeai